feat : 로또티켓
2020. 03. 17.
로또 티켓 발급

Resolves : loop-study
See also : yongdae
