+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,0 +1,7 @@
+package lotto.domain;
+
+public class LottoTicket {
+    private static final int MIN_NUMBER = 1;
+    private static final int MAX_NUMBER = 45;
+
+}
