+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,0 +1,19 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class LottoTicketTest {
+
+    @Test
+    @DisplayName("로또 발급")
+    public void createLottoTicket() throws Exception {
+        //given
+        LottoTicket lottoTicket = new LottoTicket();
+
+        //when
+
+        //then
+
+    }
+}
